+++ UOP1_Project/Assets/Art/Characters/PigChef/Animation/CaneHit.anim
@@ -75675,15 +75675,15 @@ AnimationClip:
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
   m_Events:
-  - time: 0.78333336
-    functionName: EnableWeapon
+  - time: 0.73333335
+    functionName: PlaySlashEffect
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
-  - time: 0.78333336
-    functionName: PlaySlash
+  - time: 0.76666665
+    functionName: EnableWeapon
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
