+++ UOP1_Project/Assets/Art/Characters/PigChef/Animation/CaneHit2.anim
@@ -80121,15 +80121,15 @@ AnimationClip:
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
   m_Events:
-  - time: 0.9166667
-    functionName: EnableWeapon
+  - time: 0.8666667
+    functionName: PlayReverseSlashEffect
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
   - time: 0.9166667
-    functionName: PlayReverseSlash
+    functionName: EnableWeapon
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
