--- UOP1_Project/Assets/Scripts/Characters/AttackEffectController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1199166b45e03c548965aef3749948a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
