+++ UOP1_Project/Assets/Scripts/Characters/StateMachine/Actions/PlayLiftoffJumpParticlesActionSO.cs
@@ -10,11 +10,11 @@ public class PlayLiftoffJumpParticlesActionSO : StateActionSO<PlayLiftoffJumpPar
 public class PlayLiftoffJumpParticlesAction : StateAction
 {
 	//Component references
-	private DustParticlesController _dustController;
+	private PlayerEffectContrioller _dustController;
 
 	public override void Awake(StateMachine stateMachine)
 	{
-		_dustController = stateMachine.GetComponent<DustParticlesController>();
+		_dustController = stateMachine.GetComponent<PlayerEffectContrioller>();
 	}
 
 	public override void OnStateEnter()
